+++ ip_cms/modules/standard/content_management/widgets/misc/video/module.php
@@ -16,6 +16,8 @@ global $site;
 
 $site->requireTemplate ('standard/content_management/widgets/misc/video/template.php');
 
+require_once(BASE_DIR.LIBRARY_DIR.'php/file/functions.php');
+
 class Module extends \Modules\standard\content_management\Widget{
 
 
@@ -115,16 +117,7 @@ class Module extends \Modules\standard\content_management\Widget{
     $new_name = $values['new_photo'];
     $ext_pos = strrpos($values['new_photo'], ".");
     if ($new_name != ""){
-      $new_name = substr($new_name, 0, $ext_pos);
-      $new_extension = substr($values['new_photo'], $ext_pos, strlen($values['new_photo']));
-      if (file_exists(VIDEO_DIR.$new_name.'.'.$new_extension)){
-        $i = 1;
-        while(file_exists(VIDEO_DIR.$new_name.'_'.$i.'.'.$new_extension)){
-          $i++;
-        }
-        $new_name .= '_'.$i;
-      }
-      $new_name .= "".$new_extension;
+      $new_name = \Library\Php\File\Functions::genUnocupiedName($new_name, BASE_DIR.VIDEO_DIR);
     }
 
 
@@ -160,8 +153,10 @@ class Module extends \Modules\standard\content_management\Widget{
     if(isset($values['new_photo']) && $values['new_photo'] != null){
       if (isset($values['existing_photo']) && $values['existing_photo'] != null){
         if (file_exists(VIDEO_DIR.$values['existing_photo'])){
-          if (!unlink(VIDEO_DIR.$values['existing_photo'])){
-            $this->set_error("Can't delete old photo.");
+          if ($values['existing_photo'] != '' && file_exists(VIDEO_DIR.$values['existing_photo'])){
+            if(!unlink(VIDEO_DIR.$values['existing_photo'])){
+              $this->set_error("Can't delete old photo.");
+            }
           }
         }
       }
@@ -169,21 +164,7 @@ class Module extends \Modules\standard\content_management\Widget{
 
       $new_name = $values['new_photo'];
       if ($new_name != ""){
-
-        $ext_pos = strrpos($values['new_photo'], ".");
-        if ($new_name != ""){
-          $new_name = substr($new_name, 0, $ext_pos);
-          $new_extension = substr($values['new_photo'], $ext_pos, strlen($values['new_photo']));
-          if (file_exists(VIDEO_DIR.$new_name.'.'.$new_extension)){
-            $i = 1;
-            while(file_exists(VIDEO_DIR.$new_name.'_'.$i.'.'.$new_extension)){
-              $i++;
-            }
-            $new_name .= '_'.$i;
-          }
-          $new_name .= "".$new_extension;
-        }
-
+        $new_name = \Library\Php\File\Functions::genUnocupiedName($new_name, BASE_DIR.VIDEO_DIR);
 
         copy(TMP_VIDEO_DIR.$values['new_photo'], VIDEO_DIR.$new_name);
       }
@@ -209,8 +190,10 @@ class Module extends \Modules\standard\content_management\Widget{
   function delete($values){
 
     if (isset($values['existing_photo']) && $values['existing_photo'] != null && file_exists(VIDEO_DIR.$values['existing_photo'])){
-      if (!unlink(VIDEO_DIR.$values['existing_photo'])){
-        $this->set_error("Can't delete photo.");
+      if ($values['existing_photo'] != '' && file_exists(VIDEO_DIR.$values['existing_photo'])){
+        if(!unlink(VIDEO_DIR.$values['existing_photo'])){
+          $this->set_error("Can't delete photo.");
+        }
       }
     }
 
@@ -235,8 +218,10 @@ class Module extends \Modules\standard\content_management\Widget{
     if ($rs && $lock = mysql_fetch_assoc($rs)){
 
 
-      if (!unlink(VIDEO_DIR.$lock['photo'])){
-        $this->set_error("Can't delete photo.");
+      if ($lock['photo'] != '' && file_exists(VIDEO_DIR.$lock['photo'])){
+        if(!unlink(VIDEO_DIR.$lock['photo'])){
+          $this->set_error("Can't delete photo.");
+        }
       }
