+++ ip_cms/modules/standard/content_management/widgets/text_photos/photo/module.php
@@ -12,7 +12,7 @@ const GROUP_KEY = 'text_photos';
 const MODULE_KEY = 'photo';
 
 
-
+require_once(BASE_DIR.LIBRARY_DIR.'php/file/functions.php');
 
 class Module extends \Modules\standard\content_management\Widget{
    
@@ -117,28 +117,12 @@ class Module extends \Modules\standard\content_management\Widget{
   function create_new_instance($values){
     $new_name = $values['new_photo'];
     if ($new_name != ""){
-      $new_name = substr($new_name, 0, strrpos($new_name, ".") );
-      if (file_exists(IMAGE_DIR.$new_name.'.jpg')){
-        $i = 1;
-        while(file_exists(IMAGE_DIR.$new_name.'_'.$i.'.jpg')){
-          $i++;
-        }
-        $new_name .= '_'.$i;
-      }
-      $new_name .= ".jpg";
+      $new_name = \Library\Php\File\Functions::genUnocupiedName($new_name, BASE_DIR.FILE_DIR);
     }
 
     $new_bigname = $values['new_bigphoto'];
     if ($new_bigname != ""){
-      $new_bigname = substr($new_bigname, 0, strrpos($new_bigname, ".") );
-      if (file_exists(IMAGE_DIR.$new_bigname.'.jpg') || $new_bigname.".jpg" == $new_name){
-        $i = 1;
-        while(file_exists(IMAGE_DIR.$new_bigname.'_'.$i.'.jpg') ||  $new_bigname.'_'.$i.'.jpg' == $new_name){
-          $i++;
-        }
-        $new_bigname .= '_'.$i;
-      }
-      $new_bigname .= ".jpg";
+      $new_bigname = \Library\Php\File\Functions::genUnocupiedName($new_bigname, BASE_DIR.IMAGE_DIR);
     }
 
 
@@ -175,45 +159,33 @@ class Module extends \Modules\standard\content_management\Widget{
     if(isset($values['new_photo']) && $values['new_photo'] != null){
       if (isset($values['existing_photo']) && $values['existing_photo'] != null){
         if (file_exists(IMAGE_DIR.$values['existing_photo'])){
-          if (!unlink(IMAGE_DIR.$values['existing_photo'])){
-            set_error("Can't delete old photo.");
+          if ($values['existing_photo'] != ''){
+            if(!unlink(IMAGE_DIR.$values['existing_photo'])){
+              set_error("Can't delete old photo.");
+            }
           }
         }
       }
 
       if (isset($values['existing_bigphoto']) && $values['existing_bigphoto'] != null){
         if (file_exists(IMAGE_DIR.$values['existing_bigphoto'])){
-          if (!unlink(IMAGE_DIR.$values['existing_bigphoto'])){
-            set_error("Can't delete old photo.");
+          if ($values['existing_photo'] != ''){
+            if(!unlink(IMAGE_DIR.$values['existing_bigphoto'])){
+              set_error("Can't delete old photo.");
+            }
           }
         }
       }
 
       $new_name = $values['new_photo'];
       if ($new_name != ""){
-        $new_name = substr($new_name, 0, strrpos($new_name, ".") );
-        if (file_exists(IMAGE_DIR.$new_name.'.jpg')){
-          $i = 1;
-          while(file_exists(IMAGE_DIR.$new_name.'_'.$i.'.jpg')){
-            $i++;
-          }
-          $new_name.'_'.$i;
-        }
-        $new_name .= '.jpg';
+        $new_name = \Library\Php\File\Functions::genUnocupiedName($new_name, BASE_DIR.IMAGE_DIR);
         copy(TMP_IMAGE_DIR.$values['new_photo'], IMAGE_DIR.$new_name);
       }
 
       $new_bigname = $values['new_bigphoto'];
       if ($new_bigname != ""){
-        $new_bigname = substr($new_bigname, 0, strrpos($new_bigname, ".") );
-        if (file_exists(IMAGE_DIR.$new_bigname.'.jpg') || $new_bigname.".jpg" == $new_name){
-          $i = 1;
-          while(file_exists(IMAGE_DIR.$new_bigname.'_'.$i.'.jpg') ||  $new_bigname.'_'.$i.'.jpg' == $new_name){
-            $i++;
-          }
-          $new_bigname .= '_'.$i;
-        }
-        $new_bigname .= ".jpg";
+        $new_bigname = \Library\Php\File\Functions::genUnocupiedName($new_bigname, BASE_DIR.IMAGE_DIR);
         copy(TMP_IMAGE_DIR.$values['new_bigphoto'], IMAGE_DIR.$new_bigname);
       }
 
@@ -238,13 +210,13 @@ class Module extends \Modules\standard\content_management\Widget{
   function delete($values){
 
     if (isset($values['existing_photo']) && $values['existing_photo'] != null && file_exists(IMAGE_DIR.$values['existing_photo'])){
-      if (!unlink(IMAGE_DIR.$values['existing_photo'])){
+      if ($values['existing_photo'] != '' && !unlink(IMAGE_DIR.$values['existing_photo'])){
         $this->set_error("Can't delete photo.");
       }
     }
 
     if (isset($values['existing_bigphoto']) && $values['existing_bigphoto'] != null && file_exists(IMAGE_DIR.$values['existing_bigphoto'])){
-      if (!unlink(IMAGE_DIR.$values['existing_bigphoto'])){
+      if ($values['existing_bigphoto'] != '' && !unlink(IMAGE_DIR.$values['existing_bigphoto'])){
         set_error("Can't delete photo.");
       }
     }
@@ -269,12 +241,16 @@ class Module extends \Modules\standard\content_management\Widget{
     if ($rs && $lock = mysql_fetch_assoc($rs)){
 
 
-      if (!unlink(IMAGE_DIR.$lock['photo'])){
-        set_error("Can't delete photo.");
+      if ($lock['photo'] != '' && file_exists(IMAGE_DIR.$lock['photo'])){
+        if(!unlink(IMAGE_DIR.$lock['photo'])){
+          set_error("Can't delete photo.");
+        }
       }
 
-      if (!unlink(IMAGE_DIR.$lock['photo_big'])){
-        set_error("Can't delete photo.");
+      if ($lock['photo_big'] != '' && file_exists(IMAGE_DIR.$lock['photo_big'])){
+        if(!unlink(IMAGE_DIR.$lock['photo_big'])){
+          set_error("Can't delete photo.");
+        }
       }
